Use monotonic time if available
Faulty previously used the system time for circuit timestamps, this
could lead to inaccuracies in case the system time travelled forwards or
backwards. Use monotonic time if available to resolve those issues.

To account for distributed systems, we do our best to also take real
time into account by calculating an offset from the current
monotonic time to realtime, then adding back that offset when retrieving
the time later. This is not a perfect system since different systems can
possibly treat monotonic time differently, leading to drift. However,
it's theoretically better than relying on realtime.

+++ lib/faulty/circuit.rb
@@ -547,7 +547,7 @@ class Faulty
     #
     # @return [Boolean] true if the cache should be refreshed
     def cache_should_refresh?(key)
-      time = options.cache.read(cache_refresh_key(key.to_s)).to_i
+      time = options.cache.read(cache_refresh_key(key.to_s)).to_f
       time + (((rand * 2) - 1) * options.cache_refresh_jitter) < Faulty.current_time
     end
